Move Init to fynix::init, remove crate root re-export
