+++ crates/fynix_elements/src/lib.rs
@@ -6,7 +6,6 @@ use alloc::string::String;
 use alloc::vec::Vec;
 
 use fynix::Fynix;
-use fynix::Init;
 use fynix::element::layout::ElementNodes;
 use fynix::element::meta::ElementMetas;
 use fynix::element::{
@@ -18,6 +17,7 @@ use fynix::imaging::record::{Glyph, Scene, replay_transformed};
 use fynix::imaging::{
     Composite, FillRef, GlyphRunRef, PaintSink, StrokeRef, kurbo,
 };
+use fynix::init::Init;
 use fynix::rectree::{Constraint, NodeContext, Size, Vec2};
 use parley::style::StyleProperty;
 use parley::{
